Move serde_urlencoded to src/
