+++ Cargo.toml
@@ -68,7 +68,6 @@ rand = "0.5"
 regex = "1.0"
 serde = "1.0"
 serde_json = "1.0"
-serde_urlencoded = { path = "./serde_urlencoded/" }
 sha1 = "0.6"
 smallvec = "0.6"
 time = "0.1"
@@ -103,6 +102,10 @@ tokio-tls = { version="0.1", optional = true }
 # openssl
 openssl = { version="0.10", optional = true }
 tokio-openssl = { version="0.2", optional = true }
+
+# forked url_encoded
+itoa = "0.4"
+dtoa = "0.4"
 
 [dev-dependencies]
 env_logger = "0.5"
