--- serde_urlencoded/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "serde_urlencoded"
-version = "0.5.2"
-authors = ["Anthony Ramine <n.oxyde@gmail.com>"]
-license = "MIT/Apache-2.0"
-repository = "https://github.com/nox/serde_urlencoded"
-documentation = "https://docs.rs/serde_urlencoded"
-description = "`x-www-form-urlencoded` meets Serde"
-categories = ["encoding", "web-programming"]
-keywords = ["serde", "serialization", "urlencoded"]
-
-[badges]
-travis-ci = {repository = "nox/serde_urlencoded"}
-
-[lib]
-test = false
-
-[dependencies]
-dtoa = "0.4.0"
-itoa = "0.4.0"
-serde = "1.0.0"
-url = "1.0.0"
-
-[dev-dependencies]
-serde_derive = "1.0"
